refactor: relocate SetAnonymousIdAction class
